pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aieos-pipeline-runner"
version = "0.0.0"
description = "AIEOS pipeline runner — compiles frozen CI/CD specs into execution plans"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Todd Linnertz" }]
requires-python = ">=3.11"
dependencies = [
    "jsonschema[format]>=4.20",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["aieos_pipeline_runner"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
